feat: OAuth2 유저와 일반 유저 JWT 토큰 모두 동작하게 설정

+++ src/main/java/gdsc/sc8/LIFTY/DTO/user/UserInfoResponseDto.java
@@ -8,11 +8,11 @@ public class UserInfoResponseDto {
 
     private String name;
     private String profileUri;
-    private Integer level;
+    private Long level;
     private Long exp;
 
     @Builder
-    public UserInfoResponseDto(String name, String profileUri, Integer level, Long exp) {
+    public UserInfoResponseDto(String name, String profileUri, Long level, Long exp) {
         this.name = name;
         this.profileUri = profileUri;
         this.level = level;
